+++ deployments-ucs/tenant/scripts/test-pim-auth.sh
@@ -4,12 +4,10 @@ set -ex
 export LOGIN_DATA=$(mktemp)
 export COOKIE_JAR=$(mktemp)
 
-# Workarround for self signed certificate
-CURL="curl -k --connect-timeout 10 --retry 5 --retry-delay 5 --retry-connrefused"
-# CURL="curl --connect-timeout 10 --retry 5 --retry-delay 5 --retry-connrefused"
+CURL="curl -k --connect-timeout 10 --retry 5 --retry-delay 5 --retry-connrefused --http1.1"
 
 echo "Testing ${TARGET}"
-timeout 500 curl --retry 300 --retry-delay 30 -k ${TARGET} > /dev/null
+timeout 500 curl --retry 300 --retry-delay 30 -k ${TARGET} --http1.1 > /dev/null
 echo -n "_target_path=&_username=${LOGIN}&_password=${PASSWORD}&_csrf_token=" > ${LOGIN_DATA}
 cat ${LOGIN_DATA}