+++ CHANGELOG.md
@@ -83,6 +83,7 @@
 - PIM-10674: Add lazy loading image on the asset manager library view
 - PIM-10696: Fix BETWEEN operator in GET assets endpoint
 - PIM-10705: Revert Imagick time limit because it is stateful
+- PIM-10698: Fix PHP Warning: Undefined array key "locale" in Tailored Import
 - PIM-10693: Improve the speed up of the process stop time on the rules
 
 ## Improvements